pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "specwiz-cli"
version = "0.1.0"
description = "Documentation generation platform with versioned rulebooks and structured prompts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SpecWiz Team"}
]
keywords = ["documentation", "ai", "cli", "generation", "prd", "user-guide", "release-notes"]

dependencies = [
    # CLI Framework
    "typer[all]==0.9.0",
    "rich==13.7.0",
    
    # LLM & AI
    "anthropic==0.7.0",
    
    # Data Processing
    "pydantic==2.5.0",
    
    # Templating
    "jinja2==3.1.2",
    
    # Git & Version Control
    "gitpython==3.1.40",
    
    # File I/O
    "pyyaml==6.0.1",
    
    # Events
    "blinker==1.7.0",
    
    # Utilities
    "python-dotenv==1.0.0",
    "pathspec==0.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-xdist==3.5.0",
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
    "pre-commit==3.5.0",
]

[project.scripts]
specwiz = "specwiz.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["specwiz"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "git.*",
    "blinker.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=specwiz --cov-report=term-missing --cov-report=html"
